backend/src/requirements.txt
asgiref==3.5.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.8.0
django-jazzmin==2.5.0
django-pdfkit==0.3.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
et-xmlfile==1.1.0
fonttools==4.50.0
html5lib==1.1
idna==3.4
kaleido==0.2.1
MarkupPy==1.14
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==24.0
pandas==2.2.1
pdfkit==0.6.0
Pillow==9.1.1
plotly==5.20.0
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.2
tablib==3.2.1
tenacity==8.2.3
tinycss2==1.2.1
tzdata==2024.1
urllib3==2.0.4
weasyprint==61.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
